requirements-web.txt
# Dependências CLI originais
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
anthropic>=0.29.0
numpy>=1.26.0
Pillow>=10.0.0

# Backend Web
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Autenticação
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
